Add files via upload
limite de permissões para funcionários, melhorias na paginação, correção ortográfica na navbar e troca de cor no botão de cadastrar carros.

+++ php/controller/controller_apagar_funcionario.php
@@ -2,21 +2,28 @@
 session_start();
 include_once("../conexao.php");
 
-$id = filter_input(INPUT_GET, 'id', FILTER_SANITIZE_NUMBER_INT);
+if ($_SESSION['funcionario'] == true){
+    $_SESSION['msg'] = "<p class='align-self center' style='color:red;'><b>Você Não Tem Permissão Suficiente</b></p>";
+    header("Location: ../../pages/admin/lista_funcionario.php");
+}else{
 
-if(!empty($id)) {
-    $query = "DELETE FROM usuarios WHERE pk_usuarios='$id'";
-    $query = mysqli_query($con, $query);
-    if(mysqli_affected_rows($con)) {
-        $_SESSION['msg'] = "<p class='verdinho-text'>FUNCIONARIO DELETADO COM SUCESSO</p>";
-        header("Location: ../../pages/admin/lista_funcionario.php");
-    } else {
-        $_SESSION['msg'] = "<p class='red-text'>ERRO! FUNCIOANRIO NÃO FOI DELETADO</p>";
+    $id = filter_input(INPUT_GET, 'id', FILTER_SANITIZE_NUMBER_INT);
+    
+    if(!empty($id)) {
+        $query = "DELETE FROM usuarios WHERE pk_usuarios='$id'";
+        $query = mysqli_query($con, $query);
+        if(mysqli_affected_rows($con)) {
+            $_SESSION['msg'] = "<p class='verdinho-text'>FUNCIONARIO DELETADO COM SUCESSO</p>";
+            header("Location: ../../pages/admin/lista_funcionario.php");
+        } else {
+            $_SESSION['msg'] = "<p class='red-text'>ERRO! FUNCIOANRIO NÃO FOI DELETADO</p>";
+            header("Location: ../../pages/admin/lista_funcionario.php");
+        }
+    } else{
+        $_SESSION['msg'] = "<p class='red-text'>Selecione um usuário</p";
         header("Location: ../../pages/admin/lista_funcionario.php");
     }
-} else{
-    $_SESSION['msg'] = "<p class='red-text'>Selecione um usuário</p";
-    header("Location: ../../pages/admin/lista_funcionario.php");
 }
+
 
 ?>
